[Performance] Update GeoFence security rules only if the DB transaction is successful
(cherry picked from commit 5bf6bea218cc720c8da0e4df62a19d7f266761e3)
